tmp/ref.sub.html

<!DOCTYPE html>
<html>
	<head>
		<meta charset="utf-8">
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script src="./resources/utils.sub.js"></script>
	</head>
	<body>
<script>
'use strict';
const URL_COMPLETE = new URL('/tmp/resources/test-frame.html', location).href;
const URL_DOMAIN = location.origin + '/';

const createIframe = async (t, policy) => {
	const iframe = document.createElement('iframe');
	const pipe = policy ? `?pipe=header(Referrer-Policy,${policy})` : '';
	iframe.src = `${URL_DOMAIN}tmp/resources/test-frame.html${pipe}`;

	await new Promise((resolve, reject) => {
		addEventListener('message', function handler (event) {
			if (event.source !== iframe.contentWindow) {
				return;
			}

			removeEventListener('message', handler);
			resolve();
		});
		document.body.appendChild(iframe);
		t.add_cleanup(() => iframe.remove());
	});

	return iframe.contentWindow;
};

async function send(win, message) {
	return new Promise((resolve) => {
		addEventListener('message', (event) => {
			resolve(event.data);
		});
		win.postMessage(message, '*');
	});
}

const dataUrl = () => `data:text/html,<meta charset="utf-8">
	<script src="${URL_DOMAIN}tmp/resources/utils.sub.js"><${''}/script>
	<script src="${URL_DOMAIN}tmp/resources/test-frame.js"><${''}/script>
	${Math.random()}`;

const objectUrl = (t) => {
	const encoder = new TextEncoder();
	const content = `<meta charset="utf-8">
	<script src="${URL_DOMAIN}tmp/resources/utils.sub.js"><${''}/script>
	<script src="${URL_DOMAIN}tmp/resources/test-frame.js"><${''}/script>
	${Math.random()}`;
	const url = URL.createObjectURL(
		new Blob([encoder.encode(content)], {type: 'text/html'})
	);

	t.add_cleanup(() => URL.revokeObjectURL(url));

	return url;
};

async function navigate(win, url) {
	return new Promise((resolve) => {
		addEventListener('message', function handler(event) {
			removeEventListener('message', handler);
			resolve();
		});
		win.location.assign(url);
	});
}

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const creator = await createIframe(t, 'origin');
	const subject = creator.document.createElement('iframe');

	creator.document.body.appendChild(subject);

	assert_equals(await readReferer(subject.contentWindow), '');
}, 'local - about:blank - nested - adoption');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const creator = await createIframe(t);
	const subject = creator.document.createElement('iframe');

	creator.document.body.appendChild(subject);
	createMeta(creator, 'referrer', 'origin');

	assert_equals(await readReferer(subject.contentWindow), '');
	assert_equals(await readReferer(creator), URL_DOMAIN);
}, 'local - about:blank - nested - change to creator');

promise_test(async (t) => {
	const creator = await createIframe(t);
	const subject = creator.document.createElement('iframe');

	creator.document.body.appendChild(subject);
	createMeta(subject.contentWindow, 'referrer', 'origin');

	assert_equals(await readReferer(subject.contentWindow), '');
	assert_equals(await readReferer(creator), URL_COMPLETE);
}, 'local - about:blank - nested - strengthening');

promise_test(async (t) => {
	const creator = await createIframe(t, 'origin');
	const subject = creator.document.createElement('iframe');

	creator.document.body.appendChild(subject);
	createMeta(subject.contentWindow, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(subject.contentWindow), '');
	assert_equals(await readReferer(creator), URL_DOMAIN);
}, 'local - about:blank - nested - weakening');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const creator = await createIframe(t, 'origin');

	const subject = creator.open('about:blank');
	t.add_cleanup(() => subject.close());

	assert_equals(await readReferer(subject), '');
}, 'local - about:blank - auxiliary - adoption');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const creator = await createIframe(t);
	const subject = creator.open('about:blank');
	t.add_cleanup(() => subject.close());

	createMeta(creator, 'referrer', 'origin');

	assert_equals(await readReferer(subject), '');
	assert_equals(await readReferer(creator), URL_DOMAIN);
}, 'local - about:blank - auxiliary - change to creator');

promise_test(async (t) => {
	const creator = await createIframe(t);
	const subject = creator.open('about:blank');
	t.add_cleanup(() => subject.close());

	createMeta(subject, 'referrer', 'origin');

	assert_equals(await readReferer(subject), '');
	assert_equals(await readReferer(creator), URL_COMPLETE);
}, 'local - about:blank - auxiliary - strengthening');

promise_test(async (t) => {
	const creator = await createIframe(t, 'origin');
	const subject = creator.open('about:blank');
	t.add_cleanup(() => subject.close());

	createMeta(subject, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(subject), '');
	assert_equals(await readReferer(creator), URL_DOMAIN);
}, 'local - about:blank - auxiliary - weakening');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const subject = await createIframe(t, 'origin');

	await navigate(subject, dataUrl());

	assert_equals(await send(subject, 'readReferer'), '');
}, 'local - data: - navigation - adoption');

promise_test(async (t) => {
	const subject = await createIframe(t);

	await navigate(subject, dataUrl());
	await send(subject, ['createMeta', 'referrer', 'origin']);
	assert_equals(await send(subject, 'readReferer'), '');

	await send(subject, 'back');

	assert_equals(await readReferer(subject), URL_COMPLETE);
}, 'local - data: - navigation - strengthening');

promise_test(async (t) => {
	const subject = await createIframe(t, 'origin');

	await navigate(subject, dataUrl());
	await send(subject, ['createMeta', 'referrer', 'unsafe-url']);
	assert_equals(await send(subject, 'readReferer'), '');
	await send(subject, 'back');

	assert_equals(await readReferer(subject), URL_DOMAIN);
}, 'local - data: - navigation - weakening');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const subject = await createIframe(t, 'origin');

	await navigate(subject, objectUrl(t));

	assert_equals(await send(subject, 'readReferer'), '');
}, 'local - blob: - navigation - adoption');

// This test is not relevant to behavior of "policy container" because
// ReferrerPolicy's effect on the test subject (a document whose domain uses a
// local scheme) cannot be observed.
promise_test(async (t) => {
	const subject = await createIframe(t, 'origin');
	const url = objectUrl(t);
	await send(subject, ['createMeta', 'referrer', 'unsafe-url']);

	await navigate(subject, url);

	assert_equals(await send(subject, 'readReferer'), '');
}, 'local - blob: - navigation - change to creator');

promise_test(async (t) => {
	const subject = await createIframe(t);

	await navigate(subject, objectUrl(t));
	await send(subject, ['createMeta', 'referrer', 'origin']);
	assert_equals(await send(subject, 'readReferer'), '');
	await send(subject, 'back');

	assert_equals(await readReferer(subject), URL_COMPLETE);
}, 'local - blob: - navigation - stengthening');

promise_test(async (t) => {
	const subject = await createIframe(t, 'origin');

	await navigate(subject, objectUrl(t));
	await send(subject, ['createMeta', 'referrer', 'unsafe-url']);
	assert_equals(await send(subject, 'readReferer'), '');
	await send(subject, 'back');

	assert_equals(await readReferer(subject), URL_DOMAIN);
}, 'local - blob: - navigation - weakening');
</script>
	</body>
</html>
